+++ projects/start-os/web/ui/src/app/services/api/mock-patch.ts
@@ -159,6 +159,7 @@ export const mockPatchData: DataModel = {
             upnp: { supported: null, at: null },
             pcpHostname: { supported: null, at: null },
           },
+          dnsUpdate: { supported: null, at: null },
           ipInfo: {
             name: 'Wired Connection 1',
             scopeId: 1,
@@ -180,6 +181,7 @@ export const mockPatchData: DataModel = {
             upnp: { supported: null, at: null },
             pcpHostname: { supported: null, at: null },
           },
+          dnsUpdate: { supported: null, at: null },
           ipInfo: {
             name: 'Wireless Connection 1',
             scopeId: 2,
@@ -204,6 +206,7 @@ export const mockPatchData: DataModel = {
             upnp: { supported: null, at: null },
             pcpHostname: { supported: null, at: null },
           },
+          dnsUpdate: { supported: null, at: null },
           ipInfo: {
             name: 'wireguard1',
             scopeId: 2,
@@ -228,6 +231,7 @@ export const mockPatchData: DataModel = {
             upnp: { supported: null, at: null },
             pcpHostname: { supported: null, at: null },
           },
+          dnsUpdate: { supported: null, at: null },
           ipInfo: {
             name: 'wireguard2',
             scopeId: 4,